refactor: Extract metadata panel store from file to have hot reload
